+++ tests/misc.js
@@ -498,7 +498,7 @@ describe('events', () => {
 
     let handled = false;
     const e = await jsonld.expand(d, {
-      handleEvent: ({event, next}) => {
+      eventHandler: ({event, next}) => {
         if(event.code === 'invalid reserved term') {
           handled = true;
         } else {
@@ -523,7 +523,7 @@ describe('events', () => {
     let handled0 = false;
     let handled1 = false;
     const e0 = await jsonld.expand(d, {
-      handleEvent: {
+      eventHandler: {
         'invalid reserved term': () => {
           handled0 = true;
         }
@@ -531,7 +531,7 @@ describe('events', () => {
     });
     // FIXME: ensure cache is being used
     const e1 = await jsonld.expand(d, {
-      handleEvent: {
+      eventHandler: {
         'invalid reserved term': () => {
           handled1 = true;
         }
@@ -557,7 +557,7 @@ describe('events', () => {
     let ranHandler1 = false;
     let handled = false;
     const e = await jsonld.expand(d, {
-      handleEvent: [
+      eventHandler: [
         ({next}) => {
           ranHandler0 = true;
           // skip to next handler
@@ -591,7 +591,7 @@ describe('events', () => {
 
     let handled = false;
     const e = await jsonld.expand(d, {
-      handleEvent: {
+      eventHandler: {
         'invalid reserved term': ({event}) => {
           assert.equal(event.details.term, '@RESERVED');
           handled = true;
@@ -618,7 +618,7 @@ describe('events', () => {
     let ranHandler3 = false;
     let handled = false;
     const e = await jsonld.expand(d, {
-      handleEvent: [
+      eventHandler: [
         ({next}) => {
           ranHandler0 = true;
           next();
@@ -682,7 +682,7 @@ describe('events', () => {
     let handledReservedValue = false;
     let handledLanguage = false;
     const e = await jsonld.expand(d, {
-      handleEvent: {
+      eventHandler: {
         'invalid reserved term': () => {
           handledReservedTerm = true;
         },
@@ -747,7 +747,7 @@ describe('events', () => {
     let handledLanguage2 = false;
     const e2 = await jsonld.fromRDF(d2, {
       rdfDirection: 'i18n-datatype',
-      handleEvent: {
+      eventHandler: {
         'invalid @language value': () => {
           handledLanguage2 = true;
         }