+++ CHANGELOG.md
@@ -28,7 +28,7 @@
 - Support for expansion and compaction of values container `"@direction"`.
 - Support for RDF transformation of `@direction` when `rdfDirection` is
   'i18n-datatype'.
-- Event handler option "`handleEvent`" to allow custom handling of warnings and
+- Event handler option `"eventHandler"` to allow custom handling of warnings and
   potentially other events in the future. Handles event replay for cached
   contexts.